refactor: restructure project files layout
improve project organization with unix-like file structure
