+++ lib/index.js
@@ -4,11 +4,11 @@ const cache        = {},
       log          = require('./utils').log,
       resolve      = require('./utils').resolve,
       min          = require('./min'),
-      vname        = require('./lib/vname'),
-      linkRewriter = require('./lib/link-rewriter'),
+      vname        = require('./vname'),
+      linkRewriter = require('./link-rewriter'),
       parser       = require('./parser')(),
-      sourcemaps   = require('./lib/sourcemaps'),
-      build        = require('./lib/build');
+      sourcemaps   = require('./sourcemaps'),
+      build        = require('./build');
 
 function checkOption(options, opt) {
   opt = options[opt];
