+++ lib/min/img.js
