+++ lib/min/index.js
