+++ lib/parser/general.js
