+++ lib/parser/js.js
