+++ lib/sourcemaps/index.js
