+++ lib/utils/filetype.js
