+++ lib/utils/resolve.js
