+++ package.json
@@ -17,9 +17,9 @@
     "node": ">=6",
     "npm": ">=5"
   },
-  "main": "index",
+  "main": "lib/index.js",
   "bin": {
-    "pakku": "cli.js"
+    "pakku": "bin/pakku.js"
   },
   "dependencies": {
     "autoprefixer": "^7.2.1",
